refactor(checker): extract array_literal context tests to sibling file to stay under 2000-LOC cap (#12797)
